compareR

Calculates descriptive statistics and performs statistical inference on two binary diagnostic tests in a single function call. Handles multiple comparisons using methods in `p.adjust()`.

Usage

compareR(
  df = NULL,
  test1 = NULL,
  test2 = NULL,
  gold = NULL,
  interpret = FALSE,
  alpha = 0.05,
  margins = FALSE,
  multi_corr = "holm",
  cc = TRUE,
  dp = 1,
  sesp = TRUE,
  ppvnpv = TRUE,
  plrnlr = TRUE,
  conf.int = "contemporary",
  test.names = c("Test 1", "Test 2"),
  ...
)

Arguments

df

A data frame or matrix with 3 columns (test1, test2, gold). Flexible coding of positive and negative results permitted.

test1

Either a vector of values for Test 1 (if df is NULL) or a string or integer value to be used for subsetting df.

test2

Either a vector of values for Test 2 (if df is NULL) or a string or integer value to be used for subsetting df.

gold

Either a vector of values for the gold standard test (if df is NULL) or a string or integer value to be used for subsetting df.

interpret

If TRUE provides a verbose plain English interpretation of the output by calling interpretR(). Defaults to FALSE.

alpha

An alpha value. Defaults to 0.05.

margins

A Boolean value indicating whether the contingency tables should have margins containing summed totals of rows and columns.

multi_corr

Method for multiple comparisons. Uses `p.adjust.methods`.

cc

A Boolean value indicating whether McNemar's test should be applied with continuity correction.

dp

Number of decimal places of output in summary tables. Defaults to 1.

sesp

A Boolean value indicating whether output should include sensitivity and specificity.

ppvnpv

A Boolean value indicating whether output should include positive and negative predictive values.

plrnlr

A Boolean value indicating whether output should include positive and negative likelihood ratios.

conf.int

A character string, either "contemporary" or "classic". Indicates whether function should use contemporary or classic statistical methods to calculate confidence intervals.

test.names

A vector of length two giving the names of the two different binary diagnostic tests. This argument is not relevant when testing a single binary diagnostic test.

...

Rarely needs to be used. Allows additional arguments to be passed to internal functions.

Value

A list object summarising all calculated descriptive and inferential statistics.

Details

Confidence intervals for prevalence, diagnostic accuracies and predictive values are calculated using the interval for binomial proportions described by Yu et al. (2014) by default. Setting conf.int = "classic" uses the Clopper-Pearson method. Confidence intervals for likelihood ratios are calculated using the methods recommended by Martín-Andrés and Álvarez-Hernández (2014). Hypothesis testing for diagnostic accuracies uses different methods depending on disease prevalence and number of participants or samples as described by Roldán-Nofuentes and Sidaty-Regad (2019). Global hypothesis testing for predictive values uses a method described by Roldán-Nofuentes et al. (2012), with subsequent individual tests (where indicated) performed using methods described by Kosinksi (2012). The methods for hypothesis testing- for likelihood ratios are taken from Roldán-Nofuentes & Luna del Castillo (2007).

An excellent summary of these methods is provided by Roldán-Nofuentes (2020) along with an open-source program (compbdt) licensed under GPL-2. This R package can be considered an extension of this work and is therefore distributed under the same license. Please consider citing Roldán-Nofuentes (2020) when you are citing this package.

References

Yu, Guo & Xu (2014) JSCS. 2014; 84:5,1022-1038 doi:10.1080/00949655.2012.738211

Clopper & Pearson (1934) Biometrika. 1934; 26,404-413 doi:10.2307/2331986

Martín Andrés & Álvarez Hernández (2014) Stat Comput. 2014; 24,65–75 doi:10.1007/s11222-012-9353-5

Roldán-Nofuentes & Sidaty-Regad (2019) JSCS. 2019; 89:14,2621-2644 doi:10.1080/00949655.2019.1628234

Roldán-Nofuentes, Luna del Castillo & Montero-Alonso (2012) Comput Stat Data Anal. 2012; 6,1161–1173. doi:10.1016/j.csda.2011.06.003

Kosinski (2012) Stat Med. 2012; 32,964-977 doi:10.1002/sim.5587

Roldán-Nofuentes, Luna del Castillo (2007) Stat Med. 2007; 26:4179–201. doi:10.1002/sim.2850

Roldán-Nofuentes (2020) BMC Med Res Methodol. 2020; 20,143 doi:10.1186/s12874-020-00988-y

Examples

# load data
df <- cfpr

# run compareR function
compareR(df,
  margins = TRUE, multi_corr = "bonf",
  test.names = c("pulm.exac", "pseudomonas")
)
#> Warning: Using default columns. Check test 1 is first column, test 2 is second
#>          column and gold standard is third column.
#>           test      metric estimate lower_ci upper_ci   p
#> 1    pulm.exac Sensitivity     71.8     70.6     73.0 ***
#> 2  pseudomonas Sensitivity      3.7      3.2      4.2    
#> 3    pulm.exac Specificity     80.6     79.6     81.5    
#> 4  pseudomonas Specificity     98.2     97.9     98.5 ***
#> 5    pulm.exac         PPV     73.0     71.8     74.2 ***
#> 6  pseudomonas         PPV     60.1     54.5     65.4    
#> 7    pulm.exac         NPV     79.6     78.7     80.6 ***
#> 8  pseudomonas         NPV     58.2     57.3     59.1    
#> 9    pulm.exac         PLR      3.7      3.5      3.9 ***
#> 10 pseudomonas         PLR      2.1      1.6      2.6    
#> 11   pulm.exac         NLR      0.3      0.3      0.4    
#> 12 pseudomonas         NLR      1.0      1.0      1.0 ***
#> 
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 ' ' 1